Reworked the way the button layers show/hide to fit in with the action bar and navbar appear/disappear.  Moved the manual/auto button so it doesn't get hit by the nav buttons.
