+++ app/build.gradle
@@ -55,3 +55,4 @@ dependencies {
     androidTestCompile 'org.easymock:easymock:2.5.2'
 }
 
+
