+++ app/src/main/res/values/colors.xml
@@ -1,0 +1,5 @@
+<resources>
+
+    <color name="black_overlay">#66000000</color>
+
+</resources>
