+++ app/src/main/res/values/eula.xml
@@ -4,8 +4,8 @@
     *  Twice as many stars (down to magnitude 5.6)&lt;br/&gt;
     *  Star data fixes&lt;br/&gt;
     *  Planets are now rendered in the correct order (e.g. Jupiter behind Venus)&lt;br/&gt;
-    *  Arabic translations (thanks xsoh on github)
-    *  Minor bug fixes&lt;br/&gt;
+    *  Arabic translations (thanks xsoh on github) &lt;br/&gt;
+    *  Minor bug fixes
     &lt;br/&gt;&lt;br/&gt;
     &lt;big&gt;&lt;b&gt;Privacy Policy&lt;/b&gt;&lt;/big&gt;&lt;br/&gt;
     This App uses Google Analytics to help analyse how people use it.
