+++ app/src/main/res/values/styles.xml
@@ -1,5 +1,22 @@
-<resources><style name="title" parent="text"><item name="android:textStyle">bold</item>
-</style>
-<style name="text"><item name="android:paddingLeft">5px</item>
-</style>
+<resources>
+
+    <style name="title" parent="text">
+        <item name="android:textStyle">bold</item>
+    </style>
+
+    <style name="text">
+        <item name="android:paddingLeft">5px</item>
+    </style>
+
+    <style name="FullscreenTheme" parent="@android:style/Theme.Holo">
+        <item name="android:actionBarStyle">@style/FullscreenActionBarStyle</item>
+        <item name="android:windowActionBarOverlay">true</item>
+        <item name="android:windowBackground">@null</item>
+        <item name="metaButtonBarStyle">?android:attr/buttonBarStyle</item>
+        <item name="metaButtonBarButtonStyle">?android:attr/buttonBarButtonStyle</item>
+    </style>
+
+    <style name="FullscreenActionBarStyle" parent="android:Widget.Holo.ActionBar">
+        <item name="android:background">@color/black_overlay</item>
+    </style>
 </resources>
